refactor: 연구&프로젝트 컴포넌트 디렉토리 리팩토링에 따른 page import 변경
