+++ src/app/portal/researches/projects/[id]/edit/page.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter } from 'next/navigation';
 import { use, useEffect, useState } from 'react';
-import { ProjectForm } from '@/components/researches/projects/project-form';
+import { ProjectForm } from '@/components/portal/researches/projects/project-form';
 import { Button } from '@/components/ui/button';
 import { ArrowLeft } from 'lucide-react';
 import { ProjectApi } from '@/generated-api/apis/ProjectApi';
